Create Chapter 4 - Analyzing Hashtags.ipynb
In this tutorial we will cover various analyses of the hashtags included in our Twitter dataset. Specifically, we will cover how to generate a frequency count for each of the hashtags included in our dataframe and to plot the results. We will also cover how to generate and explore summary statistics for a variable. Along the way, you will be introduced to lists and functions and for loops, among other things. Finally, we will cover how to create tag clouds.